Adding requirement of consecutive readings for a station color before stopping.
Counter added for consecutive color matches for a station, this will help with dealing with false positives readings from the Lego sensor in-between stations for station colors. Default is 3 which can be changed in the code, this small change has made my train now reach 99.99% reliability when it comes to heading to the correct station.